fix: Correct header margins for Android
